Added automatic ViewPort Zoom
Previously some screens would not display the content correctly. Either the content would be off of the screen or the content would be too small on a bigger screen.I did not want to use scroll bars so I implemented a jQuery method to ensure the viewport stays the same across computers.